Move files to better locations
